factor out ratingsCache
